third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFEComponentTransferElement-svgdom-exponent-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'exponent' property of the SVGFEComponentTransferElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS feRFunc.exponent.baseVal is 2
PASS feGFunc.exponent.baseVal is 2
PASS feBFunc.exponent.baseVal is 2
PASS feAFunc.exponent.baseVal is 2
PASS feRFunc.exponent.baseVal is 1
PASS feGFunc.exponent.baseVal is 1
PASS feBFunc.exponent.baseVal is 1
PASS feAFunc.exponent.baseVal is 1
PASS successfullyParsed is true

TEST COMPLETE
